requirements.txt
flask==3.0.0
openai>=1.40.0,<2.0.0
python-dotenv==1.0.0
Pillow==10.2.0
requests==2.31.0
